refactor(scripts): #72 2.2-6 — extract ISM/UMich/ConfBoard/NAR planner+runner blocks
Move four HTML-scrape source blocks to dedicated modules:

  scripts/validate/ism.py         196 LOC
  scripts/validate/umich.py       204 LOC
  scripts/validate/conf_board.py  229 LOC
  scripts/validate/nar.py         204 LOC

Each module owns its dataclass (`ISMProbe`, `UMichProbe`,
`ConferenceBoardProbe`, `NARProbe`), planner, schedule + current
sub-runners, and top-level `run_xxx_probe`. The validator drops
3,406 → 2,024 LOC (−772 across slices 2.2-5 and 2.2-6).

These four sources share a "schedule + current value" probe shape
(release-calendar HTML/JSON parse plus a one-shot value parse from
the indicator's current page), so each module mirrors the same
internal layout: `_run_xxx_schedule_probe`, `_run_xxx_current_probe`,
public `run_xxx_probe`.

Tests: 9 scaffold tests at 135 passed; broader subset at 695 passed.
Manual `--provider {ism,umich,conference-board,nar}` dry-runs all
plan correctly.

Refs #72.
