+++ scripts/validate_calendar_acquisition.py
@@ -109,56 +109,28 @@ from scripts.validate.census import (  # noqa: E402
     run_census_probe,
 )
 
+from scripts.validate.ism import (  # noqa: E402
+    ISMProbe,
+    plan_ism_probes,
+    run_ism_probe,
+)
+from scripts.validate.umich import (  # noqa: E402
+    UMichProbe,
+    plan_umich_probes,
+    run_umich_probe,
+)
+from scripts.validate.conf_board import (  # noqa: E402
+    ConferenceBoardProbe,
+    plan_conference_board_probes,
+    run_conference_board_probe,
+)
+from scripts.validate.nar import (  # noqa: E402
+    NARProbe,
+    plan_nar_probes,
+    run_nar_probe,
+)
+
 # Scaffold under test — we import and exercise exactly what production uses.
-from ingestion.calendar.ism_api import (  # noqa: E402
-    ISM_RELEASE_CALENDAR_URL,
-    ISM_REPORTS_URL,
-    discover_current_report_url,
-    fetch_report_html as fetch_ism_report_html,
-    fetch_reports_landing_html as fetch_ism_reports_landing_html,
-    fetch_schedule_html as fetch_ism_schedule_html,
-    parse_report_html as parse_ism_report_html,
-    parse_schedule_html as parse_ism_schedule_html,
-    report_value_to_records as ism_report_value_to_records,
-    schedule_entry_to_records as ism_schedule_entry_to_records,
-)
-
-from ingestion.calendar.umich_api import (  # noqa: E402
-    UMICH_MAIN_URL,
-    UMICH_SURVEY_INFO_URL,
-    UMichScheduleDocument,
-    current_value_to_records as umich_current_value_to_records,
-    fetch_current_results_html as fetch_umich_current_results_html,
-    fetch_release_dates_document as fetch_umich_release_dates_document,
-    parse_current_results_html as parse_umich_current_results_html,
-    parse_release_dates_text as parse_umich_release_dates_text,
-    schedule_entry_to_records as umich_schedule_entry_to_records,
-)
-
-from ingestion.calendar.conference_board_api import (  # noqa: E402
-    CONFERENCE_BOARD_CALENDAR_URL,
-    CONFERENCE_BOARD_CONSUMER_CONFIDENCE_URL,
-    CONFERENCE_BOARD_LEADING_INDICATORS_URL,
-    current_value_to_records as conference_board_current_value_to_records,
-    fetch_calendar_json as fetch_conference_board_calendar_json,
-    fetch_indicator_html as fetch_conference_board_indicator_html,
-    parse_calendar_events_json as parse_conference_board_calendar_events_json,
-    parse_current_value_html as parse_conference_board_current_value_html,
-    schedule_entry_to_records as conference_board_schedule_entry_to_records,
-)
-
-from ingestion.calendar.nar_api import (  # noqa: E402
-    NAR_EXISTING_HOME_SALES_URL,
-    NAR_PENDING_HOME_SALES_URL,
-    NAR_SCHEDULE_URL,
-    current_value_to_records as nar_current_value_to_records,
-    fetch_current_html as fetch_nar_current_html,
-    fetch_schedule_html as fetch_nar_schedule_html,
-    parse_current_value_html as parse_nar_current_value_html,
-    parse_schedule_html as parse_nar_schedule_html,
-    schedule_entry_to_records as nar_schedule_entry_to_records,
-)
-
 from ingestion.calendar.ecb_api import (  # noqa: E402
     INDICATOR_REGISTRY as ECB_INDICATOR_REGISTRY,
     parse_observation as parse_ecb_observation,
@@ -401,750 +373,6 @@ def _fmt_field_list(fields: list[str]) -> str:
         return "(none)"
     return ", ".join(f"`{f}`" for f in fields)
 
-
-
-# ──────────────────────────────────────────────────────────────────────────
-# ISM probe plan + runner
-# ──────────────────────────────────────────────────────────────────────────
-
-
-@dataclass
-class ISMProbe:
-    """One ISM live-validation probe."""
-
-    name: str
-    source: str
-    url: str
-    description: str
-
-
-def plan_ism_probes() -> list[ISMProbe]:
-    """Validate the ISM release calendar and current Manufacturing report."""
-    return [
-        ISMProbe(
-            name="ism_release_calendar",
-            source="schedule",
-            url=ISM_RELEASE_CALENDAR_URL,
-            description=(
-                "ISM Manufacturing PMI release dates from the public "
-                "release-calendar table"
-            ),
-        ),
-        ISMProbe(
-            name="ism_current_manufacturing_report",
-            source="report",
-            url=ISM_REPORTS_URL,
-            description=(
-                "ISM PMI reports hub discovery plus current Manufacturing "
-                "PMI report value parse"
-            ),
-        ),
-    ]
-
-
-def _try_project_ism_schedule(entry: Any) -> tuple[bool, str]:
-    try:
-        raw, event = ism_schedule_entry_to_records(
-            entry,
-            snapshot_epoch_ms=1_700_000_000_000,
-        )
-        return True, (
-            f"ok series={entry.series_id} ref={event.reference_date} "
-            f"event_id={raw.provider_event_id[:10]}..."
-        )
-    except Exception as exc:
-        return False, f"{type(exc).__name__}: {exc}"
-
-
-def _try_project_ism_value(value: Any) -> tuple[bool, str]:
-    try:
-        raw, event = ism_report_value_to_records(
-            value,
-            snapshot_epoch_ms=1_700_000_000_000,
-        )
-        return True, (
-            f"ok series={value.series_id} actual={event.actual} "
-            f"event_id={raw.provider_event_id[:10]}..."
-        )
-    except Exception as exc:
-        return False, f"{type(exc).__name__}: {exc}"
-
-
-def _run_ism_schedule_probe(
-    result: ProbeResult,
-    schedule_fetcher,
-) -> ProbeResult:
-    import time as _time
-
-    t0 = _time.monotonic()
-    try:
-        html = schedule_fetcher()
-        entries = parse_ism_schedule_html(html)
-    except Exception as exc:
-        result.status = "http_error"
-        result.notes.append(f"{type(exc).__name__}: {exc}")
-        return result
-    result.http_elapsed_ms = (_time.monotonic() - t0) * 1000
-    result.row_count = len(entries)
-    if not entries:
-        result.status = "http_error"
-        result.notes.append("zero ISM schedule entries parsed")
-        return result
-    result.status = "ok"
-    ordered = sorted(entries, key=lambda e: e.release_date, reverse=True)
-    sample = ordered[0]
-    result.sample_row = {
-        "series_id": sample.series_id,
-        "reference_date": sample.reference_date,
-        "release_date": sample.release_date,
-        "event_time_utc": sample.event_time_utc,
-    }
-    by_series = Counter(entry.series_id for entry in entries)
-    result.enum_counters = {"series_id": by_series}
-    result.notes.append(
-        "entries by series: "
-        + ", ".join(f"{k}={v}" for k, v in by_series.most_common())
-    )
-    sample_n = min(10, len(ordered))
-    result.parse_attempts = sample_n
-    for entry in ordered[:sample_n]:
-        ok, msg = _try_project_ism_schedule(entry)
-        if ok:
-            result.parse_successes += 1
-        else:
-            if len(result.parse_error_samples) < 3:
-                result.parse_error_samples.append(msg)
-    return result
-
-
-def _run_ism_report_probe(
-    result: ProbeResult,
-    landing_fetcher,
-    report_fetcher,
-) -> ProbeResult:
-    import time as _time
-
-    t0 = _time.monotonic()
-    try:
-        landing_html = landing_fetcher()
-        report_url = discover_current_report_url(landing_html)
-        result.request_path = f"GET {ISM_REPORTS_URL} -> GET {report_url}"
-        report_html = report_fetcher(report_url)
-        value = parse_ism_report_html(report_html, source_url=report_url)
-    except Exception as exc:
-        result.status = "http_error"
-        result.notes.append(f"{type(exc).__name__}: {exc}")
-        return result
-    result.http_elapsed_ms = (_time.monotonic() - t0) * 1000
-    result.status = "ok"
-    result.row_count = 1
-    result.sample_row = {
-        "series_id": value.series_id,
-        "reference_date": value.reference_date,
-        "actual": value.actual,
-        "previous": value.previous,
-        "source_url": value.source_url,
-    }
-    result.parse_attempts = 1
-    ok, msg = _try_project_ism_value(value)
-    if ok:
-        result.parse_successes = 1
-    else:
-        result.parse_error_samples.append(msg)
-    return result
-
-
-def run_ism_probe(
-    probe: ISMProbe,
-    *,
-    schedule_fetcher=fetch_ism_schedule_html,
-    landing_fetcher=fetch_ism_reports_landing_html,
-    report_fetcher=fetch_ism_report_html,
-) -> ProbeResult:
-    """Execute one ISM public-HTML probe."""
-    generic = Probe(
-        name=probe.name,
-        path=f"GET {probe.url}",
-        description=probe.description,
-        expected_shape="HTML -> ISM schedule entries / report value",
-        expected_fields=frozenset(),
-    )
-    result = ProbeResult(probe=generic, status="skipped")
-    result.request_path = generic.path
-    if probe.source == "schedule":
-        return _run_ism_schedule_probe(result, schedule_fetcher)
-    if probe.source == "report":
-        return _run_ism_report_probe(result, landing_fetcher, report_fetcher)
-    raise ValueError(f"unknown ISM probe source: {probe.source!r}")
-
-
-# ──────────────────────────────────────────────────────────────────────────
-# U Michigan probe plan + runner
-# ──────────────────────────────────────────────────────────────────────────
-
-
-@dataclass
-class UMichProbe:
-    """One U Michigan live-validation probe."""
-
-    name: str
-    source: str
-    url: str
-    description: str
-    year: int | None = None
-
-
-def plan_umich_probes() -> list[UMichProbe]:
-    """Validate the U Michigan release-date PDF and current results page."""
-    year = datetime.now(timezone.utc).year
-    return [
-        UMichProbe(
-            name=f"umich_release_dates_{year}",
-            source="schedule",
-            url=UMICH_SURVEY_INFO_URL,
-            description=(
-                "U Michigan Consumer Sentiment preliminary/final release "
-                f"dates for {year}"
-            ),
-            year=year,
-        ),
-        UMichProbe(
-            name="umich_current_results",
-            source="current",
-            url=UMICH_MAIN_URL,
-            description=(
-                "U Michigan current Consumer Sentiment table value parse"
-            ),
-        ),
-    ]
-
-
-def _try_project_umich_schedule(entry: Any) -> tuple[bool, str]:
-    try:
-        raw, event = umich_schedule_entry_to_records(
-            entry,
-            snapshot_epoch_ms=1_700_000_000_000,
-        )
-        return True, (
-            f"ok series={entry.series_id} stage={entry.release_stage} "
-            f"ref={event.reference_date} event_id={raw.provider_event_id[:10]}..."
-        )
-    except Exception as exc:
-        return False, f"{type(exc).__name__}: {exc}"
-
-
-def _try_project_umich_value(value: Any) -> tuple[bool, str]:
-    try:
-        raw, event = umich_current_value_to_records(
-            value,
-            snapshot_epoch_ms=1_700_000_000_000,
-        )
-        return True, (
-            f"ok series={value.series_id} stage={value.release_stage} "
-            f"actual={event.actual} event_id={raw.provider_event_id[:10]}..."
-        )
-    except Exception as exc:
-        return False, f"{type(exc).__name__}: {exc}"
-
-
-def _coerce_umich_document(doc: object) -> tuple[str, str]:
-    if isinstance(doc, UMichScheduleDocument):
-        return doc.text, doc.source_url
-    return str(doc), UMICH_SURVEY_INFO_URL
-
-
-def _run_umich_schedule_probe(
-    result: ProbeResult,
-    probe: UMichProbe,
-    document_fetcher,
-) -> ProbeResult:
-    import time as _time
-
-    t0 = _time.monotonic()
-    try:
-        doc = document_fetcher(year=probe.year)
-        text, source_url = _coerce_umich_document(doc)
-        result.request_path = f"GET {UMICH_SURVEY_INFO_URL} -> GET {source_url}"
-        entries = parse_umich_release_dates_text(text, source_url=source_url)
-    except Exception as exc:
-        result.status = "http_error"
-        result.notes.append(f"{type(exc).__name__}: {exc}")
-        return result
-    result.http_elapsed_ms = (_time.monotonic() - t0) * 1000
-    result.row_count = len(entries)
-    if not entries:
-        result.status = "http_error"
-        result.notes.append("zero U Michigan schedule entries parsed")
-        return result
-    result.status = "ok"
-    ordered = sorted(entries, key=lambda e: e.release_date, reverse=True)
-    sample = ordered[0]
-    result.sample_row = {
-        "series_id": sample.series_id,
-        "reference_date": sample.reference_date,
-        "release_stage": sample.release_stage,
-        "release_date": sample.release_date,
-        "event_time_utc": sample.event_time_utc,
-    }
-    by_stage = Counter(entry.release_stage for entry in entries)
-    result.enum_counters = {"release_stage": by_stage}
-    result.notes.append(
-        "entries by stage: "
-        + ", ".join(f"{k}={v}" for k, v in by_stage.most_common())
-    )
-    sample_n = min(10, len(ordered))
-    result.parse_attempts = sample_n
-    for entry in ordered[:sample_n]:
-        ok, msg = _try_project_umich_schedule(entry)
-        if ok:
-            result.parse_successes += 1
-        else:
-            if len(result.parse_error_samples) < 3:
-                result.parse_error_samples.append(msg)
-    return result
-
-
-def _run_umich_current_probe(
-    result: ProbeResult,
-    current_fetcher,
-) -> ProbeResult:
-    import time as _time
-
-    t0 = _time.monotonic()
-    try:
-        html = current_fetcher()
-        value = parse_umich_current_results_html(html, source_url=UMICH_MAIN_URL)
-    except Exception as exc:
-        result.status = "http_error"
-        result.notes.append(f"{type(exc).__name__}: {exc}")
-        return result
-    result.http_elapsed_ms = (_time.monotonic() - t0) * 1000
-    result.status = "ok"
-    result.row_count = 1
-    result.sample_row = {
-        "series_id": value.series_id,
-        "reference_date": value.reference_date,
-        "release_stage": value.release_stage,
-        "actual": value.actual,
-        "previous": value.previous,
-        "source_url": value.source_url,
-    }
-    result.parse_attempts = 1
-    ok, msg = _try_project_umich_value(value)
-    if ok:
-        result.parse_successes = 1
-    else:
-        result.parse_error_samples.append(msg)
-    return result
-
-
-def run_umich_probe(
-    probe: UMichProbe,
-    *,
-    document_fetcher=fetch_umich_release_dates_document,
-    current_fetcher=fetch_umich_current_results_html,
-) -> ProbeResult:
-    """Execute one U Michigan public HTML/PDF probe."""
-    generic = Probe(
-        name=probe.name,
-        path=f"GET {probe.url}",
-        description=probe.description,
-        expected_shape="HTML/PDF -> U Michigan schedule entries / current value",
-        expected_fields=frozenset(),
-    )
-    result = ProbeResult(probe=generic, status="skipped")
-    result.request_path = generic.path
-    if probe.source == "schedule":
-        return _run_umich_schedule_probe(result, probe, document_fetcher)
-    if probe.source == "current":
-        return _run_umich_current_probe(result, current_fetcher)
-    raise ValueError(f"unknown U Michigan probe source: {probe.source!r}")
-
-
-# ──────────────────────────────────────────────────────────────────────────
-# Conference Board probe plan + runner
-# ──────────────────────────────────────────────────────────────────────────
-
-
-@dataclass
-class ConferenceBoardProbe:
-    """One Conference Board live-validation probe."""
-
-    name: str
-    source: str
-    url: str
-    description: str
-    series_id: str = ""
-    from_epoch_ms: int | None = None
-    to_epoch_ms: int | None = None
-
-
-def _conference_board_window() -> tuple[int, int]:
-    today = datetime.now(timezone.utc).date()
-    start = today - timedelta(days=45)
-    end = today + timedelta(days=180)
-    start_ms = int(datetime(start.year, start.month, start.day, tzinfo=timezone.utc).timestamp() * 1000)
-    end_ms = int(datetime(end.year, end.month, end.day, tzinfo=timezone.utc).timestamp() * 1000)
-    return start_ms, end_ms
-
-
-def plan_conference_board_probes() -> list[ConferenceBoardProbe]:
-    """Validate the Conference Board calendar endpoint and current pages."""
-    from_ms, to_ms = _conference_board_window()
-    return [
-        ConferenceBoardProbe(
-            name="conference_board_release_calendar",
-            source="schedule",
-            url=CONFERENCE_BOARD_CALENDAR_URL,
-            description=(
-                "Conference Board economic-indicator calendar rows for "
-                "US Consumer Confidence and US Leading Index"
-            ),
-            from_epoch_ms=from_ms,
-            to_epoch_ms=to_ms,
-        ),
-        ConferenceBoardProbe(
-            name="conference_board_consumer_confidence",
-            source="current",
-            url=CONFERENCE_BOARD_CONSUMER_CONFIDENCE_URL,
-            description="Conference Board current Consumer Confidence value parse",
-            series_id="TCB_CONSUMER_CONFIDENCE",
-        ),
-        ConferenceBoardProbe(
-            name="conference_board_us_leading_index",
-            source="current",
-            url=CONFERENCE_BOARD_LEADING_INDICATORS_URL,
-            description="Conference Board current US Leading Index monthly-change parse",
-            series_id="TCB_LEADING_INDEX",
-        ),
-    ]
-
-
-def _try_project_conference_board_schedule(entry: Any) -> tuple[bool, str]:
-    try:
-        raw, event = conference_board_schedule_entry_to_records(
-            entry,
-            snapshot_epoch_ms=1_700_000_000_000,
-        )
-        return True, (
-            f"ok series={entry.series_id} ref={event.reference_date} "
-            f"event_id={raw.provider_event_id[:10]}..."
-        )
-    except Exception as exc:
-        return False, f"{type(exc).__name__}: {exc}"
-
-
-def _try_project_conference_board_value(value: Any) -> tuple[bool, str]:
-    try:
-        raw, event = conference_board_current_value_to_records(
-            value,
-            snapshot_epoch_ms=1_700_000_000_000,
-        )
-        return True, (
-            f"ok series={value.series_id} actual={event.actual} "
-            f"event_id={raw.provider_event_id[:10]}..."
-        )
-    except Exception as exc:
-        return False, f"{type(exc).__name__}: {exc}"
-
-
-def _run_conference_board_schedule_probe(
-    result: ProbeResult,
-    probe: ConferenceBoardProbe,
-    schedule_fetcher,
-) -> ProbeResult:
-    import time as _time
-
-    from_ms, to_ms = probe.from_epoch_ms, probe.to_epoch_ms
-    if from_ms is None or to_ms is None:
-        from_ms, to_ms = _conference_board_window()
-    result.request_path = f"GET {probe.url}?from={from_ms}&to={to_ms}"
-    t0 = _time.monotonic()
-    try:
-        payload = schedule_fetcher(from_epoch_ms=from_ms, to_epoch_ms=to_ms)
-        entries = parse_conference_board_calendar_events_json(payload)
-    except Exception as exc:
-        result.status = "http_error"
-        result.notes.append(f"{type(exc).__name__}: {exc}")
-        return result
-    result.http_elapsed_ms = (_time.monotonic() - t0) * 1000
-    result.row_count = len(entries)
-    if not entries:
-        result.status = "http_error"
-        result.notes.append("zero Conference Board schedule entries parsed")
-        return result
-    result.status = "ok"
-    ordered = sorted(entries, key=lambda e: e.release_date, reverse=True)
-    sample = ordered[0]
-    result.sample_row = {
-        "series_id": sample.series_id,
-        "calendar_event_id": sample.calendar_event_id,
-        "reference_date": sample.reference_date,
-        "release_date": sample.release_date,
-        "event_time_utc": sample.event_time_utc,
-        "source_url": sample.source_url,
-    }
-    by_series = Counter(entry.series_id for entry in entries)
-    result.enum_counters = {"series_id": by_series}
-    result.notes.append(
-        "entries by series: "
-        + ", ".join(f"{k}={v}" for k, v in by_series.most_common())
-    )
-    sample_n = min(10, len(ordered))
-    result.parse_attempts = sample_n
-    for entry in ordered[:sample_n]:
-        ok, msg = _try_project_conference_board_schedule(entry)
-        if ok:
-            result.parse_successes += 1
-        else:
-            if len(result.parse_error_samples) < 3:
-                result.parse_error_samples.append(msg)
-    return result
-
-
-def _run_conference_board_current_probe(
-    result: ProbeResult,
-    probe: ConferenceBoardProbe,
-    current_fetcher,
-) -> ProbeResult:
-    import time as _time
-
-    t0 = _time.monotonic()
-    try:
-        html = current_fetcher(probe.url)
-        value = parse_conference_board_current_value_html(
-            html,
-            source_url=probe.url,
-            series_id=probe.series_id,
-        )
-    except Exception as exc:
-        result.status = "http_error"
-        result.notes.append(f"{type(exc).__name__}: {exc}")
-        return result
-    result.http_elapsed_ms = (_time.monotonic() - t0) * 1000
-    result.status = "ok"
-    result.row_count = 1
-    result.sample_row = {
-        "series_id": value.series_id,
-        "reference_date": value.reference_date,
-        "actual": value.actual,
-        "previous": value.previous,
-        "index_level": value.index_level,
-        "source_url": value.source_url,
-    }
-    result.parse_attempts = 1
-    ok, msg = _try_project_conference_board_value(value)
-    if ok:
-        result.parse_successes = 1
-    else:
-        result.parse_error_samples.append(msg)
-    return result
-
-
-def run_conference_board_probe(
-    probe: ConferenceBoardProbe,
-    *,
-    schedule_fetcher=fetch_conference_board_calendar_json,
-    current_fetcher=fetch_conference_board_indicator_html,
-) -> ProbeResult:
-    """Execute one Conference Board public HTML/JSON probe."""
-    generic = Probe(
-        name=probe.name,
-        path=f"GET {probe.url}",
-        description=probe.description,
-        expected_shape="JSON/HTML -> Conference Board schedule entries / current values",
-        expected_fields=frozenset(),
-    )
-    result = ProbeResult(probe=generic, status="skipped")
-    result.request_path = generic.path
-    if probe.source == "schedule":
-        return _run_conference_board_schedule_probe(result, probe, schedule_fetcher)
-    if probe.source == "current":
-        return _run_conference_board_current_probe(result, probe, current_fetcher)
-    raise ValueError(f"unknown Conference Board probe source: {probe.source!r}")
-
-
-# ──────────────────────────────────────────────────────────────────────────
-# NAR probe plan + runner
-# ──────────────────────────────────────────────────────────────────────────
-
-
-@dataclass
-class NARProbe:
-    """One NAR live-validation probe."""
-
-    name: str
-    source: str
-    url: str
-    description: str
-    series_id: str = ""
-
-
-def plan_nar_probes() -> list[NARProbe]:
-    """Validate the NAR schedule page and current housing-statistics pages."""
-    return [
-        NARProbe(
-            name="nar_statistical_release_schedule",
-            source="schedule",
-            url=NAR_SCHEDULE_URL,
-            description=(
-                "NAR statistical release schedule rows for Existing-Home "
-                "Sales and Pending Home Sales Index"
-            ),
-        ),
-        NARProbe(
-            name="nar_existing_home_sales",
-            source="current",
-            url=NAR_EXISTING_HOME_SALES_URL,
-            description="NAR current Existing Home Sales million-SAAR parse",
-            series_id="NAR_EXISTING_HOME_SALES",
-        ),
-        NARProbe(
-            name="nar_pending_home_sales",
-            source="current",
-            url=NAR_PENDING_HOME_SALES_URL,
-            description="NAR current Pending Home Sales MoM percent parse",
-            series_id="NAR_PENDING_HOME_SALES_MOM",
-        ),
-    ]
-
-
-def _try_project_nar_schedule(entry: Any) -> tuple[bool, str]:
-    try:
-        raw, event = nar_schedule_entry_to_records(
-            entry,
-            snapshot_epoch_ms=1_700_000_000_000,
-        )
-        return True, (
-            f"ok series={entry.series_id} ref={event.reference_date} "
-            f"event_id={raw.provider_event_id[:10]}..."
-        )
-    except Exception as exc:
-        return False, f"{type(exc).__name__}: {exc}"
-
-
-def _try_project_nar_value(value: Any) -> tuple[bool, str]:
-    try:
-        raw, event = nar_current_value_to_records(
-            value,
-            snapshot_epoch_ms=1_700_000_000_000,
-        )
-        return True, (
-            f"ok series={value.series_id} actual={event.actual} "
-            f"event_id={raw.provider_event_id[:10]}..."
-        )
-    except Exception as exc:
-        return False, f"{type(exc).__name__}: {exc}"
-
-
-def _run_nar_schedule_probe(
-    result: ProbeResult,
-    schedule_fetcher,
-) -> ProbeResult:
-    import time as _time
-
-    t0 = _time.monotonic()
-    try:
-        html = schedule_fetcher()
-        entries = parse_nar_schedule_html(html)
-    except Exception as exc:
-        result.status = "http_error"
-        result.notes.append(f"{type(exc).__name__}: {exc}")
-        return result
-    result.http_elapsed_ms = (_time.monotonic() - t0) * 1000
-    result.row_count = len(entries)
-    if not entries:
-        result.status = "http_error"
-        result.notes.append("zero NAR schedule entries parsed")
-        return result
-    result.status = "ok"
-    ordered = sorted(entries, key=lambda e: e.release_date, reverse=True)
-    sample = ordered[0]
-    result.sample_row = {
-        "series_id": sample.series_id,
-        "raw_title": sample.raw_title,
-        "reference_date": sample.reference_date,
-        "release_date": sample.release_date,
-        "event_time_utc": sample.event_time_utc,
-    }
-    by_series = Counter(entry.series_id for entry in entries)
-    result.enum_counters = {"series_id": by_series}
-    result.notes.append(
-        "entries by series: "
-        + ", ".join(f"{k}={v}" for k, v in by_series.most_common())
-    )
-    sample_n = min(10, len(ordered))
-    result.parse_attempts = sample_n
-    for entry in ordered[:sample_n]:
-        ok, msg = _try_project_nar_schedule(entry)
-        if ok:
-            result.parse_successes += 1
-        else:
-            if len(result.parse_error_samples) < 3:
-                result.parse_error_samples.append(msg)
-    return result
-
-
-def _run_nar_current_probe(
-    result: ProbeResult,
-    probe: NARProbe,
-    current_fetcher,
-) -> ProbeResult:
-    import time as _time
-
-    t0 = _time.monotonic()
-    try:
-        html = current_fetcher(probe.url)
-        value = parse_nar_current_value_html(
-            html,
-            source_url=probe.url,
-            series_id=probe.series_id,
-        )
-    except Exception as exc:
-        result.status = "http_error"
-        result.notes.append(f"{type(exc).__name__}: {exc}")
-        return result
-    result.http_elapsed_ms = (_time.monotonic() - t0) * 1000
-    result.status = "ok"
-    result.row_count = 1
-    result.sample_row = {
-        "series_id": value.series_id,
-        "reference_date": value.reference_date,
-        "actual": value.actual,
-        "previous": value.previous,
-        "raw_change": value.raw_change,
-        "source_url": value.source_url,
-    }
-    result.parse_attempts = 1
-    ok, msg = _try_project_nar_value(value)
-    if ok:
-        result.parse_successes = 1
-    else:
-        result.parse_error_samples.append(msg)
-    return result
-
-
-def run_nar_probe(
-    probe: NARProbe,
-    *,
-    schedule_fetcher=fetch_nar_schedule_html,
-    current_fetcher=fetch_nar_current_html,
-) -> ProbeResult:
-    """Execute one NAR public HTML probe."""
-    generic = Probe(
-        name=probe.name,
-        path=f"GET {probe.url}",
-        description=probe.description,
-        expected_shape="HTML -> NAR schedule entries / current values",
-        expected_fields=frozenset(),
-    )
-    result = ProbeResult(probe=generic, status="skipped")
-    result.request_path = generic.path
-    if probe.source == "schedule":
-        return _run_nar_schedule_probe(result, schedule_fetcher)
-    if probe.source == "current":
-        return _run_nar_current_probe(result, probe, current_fetcher)
-    raise ValueError(f"unknown NAR probe source: {probe.source!r}")
 
 
 # ──────────────────────────────────────────────────────────────────────────
